requirements.txt
streamlit
snowflake-connector-python
pandas
plotly
numpy